+++ JudoPayDotNet/Models/WebPaymentResponseModel.cs
@@ -12,19 +12,24 @@ namespace JudoPayDotNet.Models
 // ReSharper restore ClassNeverInstantiated.Global
     {
         /// <summary>
-        /// Gets or sets the post URL.
+        /// URL prefix to launch Judo hosted web payments page (follow with ?reference={Reference})
         /// </summary>
-        /// <value>
-        /// The post URL.
-        /// </value>
         [DataMember(EmitDefaultValue = false)]
-        public string PostUrl { get; set; }
+        public string PostUrl { get; private set; }
 
         /// <summary>
         /// Judopay generated reference for the web payment.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public string Reference { get; set; }
+        public string Reference { get; private set; }
+
+        [DataMember(EmitDefaultValue = false)]
+        private string PayByLinkUrl { get; set; }
+
+        /// <summary>
+        /// Short URL to launch Judo hosted web payments page
+        /// </summary>
+        public string ShortUrl => PayByLinkUrl;
     }
     // ReSharper restore UnusedAutoPropertyAccessor.Global
 }